Module type Cont.S

The unary monad interface

include Trans.S

type 'a t

type 'a m

type 'a e

val lift : 'a m -> 'a t

lifts inner monad into the resulting monad

val run : 'a t -> 'a e

runs the computation

include Monad with type 'a t := 'a t

val void : 'a t -> unit t

void m computes m and discrards the result.

val sequence : unit t list -> unit t

sequence xs computes a sequence of computations xs in the left to right order.

val forever : 'a t -> 'b t

forever xs creates a computationt that never returns.

module Fn : sig ... end

Various function combinators lifted into the Kleisli category.

module Pair : sig ... end

The pair interface lifted into the monad.

module Triple : sig ... end

The triple interface lifted into a monad.

module Lift : sig ... end

Lifts functions into the monad.

module Exn : sig ... end

Interacting between monads and language exceptions

module Collection : sig ... end

Lifts collection interface into the monad.

module List : Collection.S with type 'a t := 'a list

The Monad.Collection.S interface for lists

module Seq : Collection.S with type 'a t := 'a Core_kernel.Sequence.t

The Monad.Collection.S interface for sequences

include Syntax.S with type 'a t := 'a t

val (>>=) : 'a t -> ( 'a -> 'b t ) -> 'b t

m >>= f is bind m f

val (>>|) : 'a t -> ( 'a -> 'b ) -> 'b t

m >>= f is map m ~f

val (>=>) : ( 'a -> 'b t ) -> ( 'b -> 'c t ) -> 'a -> 'c t

f >=> g is fun x -> f x >>= g

val (!!) : 'a -> 'a t

!!x is return x

val (!$) : ( 'a -> 'b ) -> 'a t -> 'b t

!$f is Lift.unary f

val (!$$) : ( 'a -> 'b -> 'c ) -> 'a t -> 'b t -> 'c t

!$$f is Lift.binary f

val (!$$$) : ( 'a -> 'b -> 'c -> 'd ) -> 'a t -> 'b t -> 'c t -> 'd t

!$$$f is Lift.ternary f

val (!$$$$) : ( 'a -> 'b -> 'c -> 'd -> 'e ) -> 'a t -> 'b t -> 'c t -> 'd t -> 'e t

!$$$$f is Lift.quaternary f

val (!$$$$$) : ( 'a -> 'b -> 'c -> 'd -> 'e -> 'f ) -> 'a t -> 'b t -> 'c t -> 'd t -> 'e t -> 'f t

!$$$$$f is Lift.quinary f

include Syntax.Let.S with type 'a t := 'a t

val let* : 'a t -> ( 'a -> 'b t ) -> 'b t

let* r = f x in b is f x >>= fun r -> b

val and* : 'a t -> 'b t -> ('a * 'b) t

monoidal product

val let+ : 'a t -> ( 'a -> 'b ) -> 'b t

let+ r = f x in b is f x >>| fun r -> b

val and+ : 'a t -> 'b t -> ('a * 'b) t

monoidal product

include Core_kernel.Monad.S with type 'a t := 'a t

val (>>=) : 'a t -> ( 'a -> 'b t ) -> 'b t

val (>>|) : 'a t -> ( 'a -> 'b ) -> 'b t

module Monad_infix : sig ... end

val bind : 'a t -> f:( 'a -> 'b t ) -> 'b t

val return : 'a -> 'a t

val map : 'a t -> f:( 'a -> 'b ) -> 'b t

val join : 'a t t -> 'a t

val ignore_m : 'a t -> unit t

val all : 'a t list -> 'a list t

val all_unit : unit t list -> unit t

module Let_syntax : sig ... end

module Let : Syntax.Let.S with type 'a t := 'a t

Monadic operators, see Monad.Syntax.S for more.

module Syntax : Syntax.S with type 'a t := 'a t

Monadic operators, see Monad.Syntax.S for more.

type r

type of the final result of computation

val call : f:( cc:( 'a -> _ t ) -> 'a t ) -> 'a t

call ~f calls f ~cc with the current continuation cc.

The call ~f computation may be computed more than once, i.e., it would be resumed every time the continuation is invoked. The captured continuation represents the computation around the call. Thus invoking this computation will effectively escape the f function (discarding the consequent computations) and continue with a computation that follows the call. The continuation is multi-shot, in the sense that it can be called (resumed) multiple times (or not called at all). Every time it is called, the computation will resume at the same point, thus a computation that contains the call can be seen as a reenterable computation, and the call itself marks the entry point, and the continuation acts like a key that allows any computation that has it to reenter the subroutine at this point.