Module Monad.Option

The option aka Maybe monad.

This monad can be used to denote partial computations or a limited form of non-deterministic computations where an absence of result, i.e., the bottom value, can be considered as a possible outcome.

module type S = sig ... end

The unary option monad.

module type S2 = sig ... end

The binary option monad

include S with type 'a S.t = 'a option and type 'a S.m = 'a and type 'a S.e = 'a option

type 'a t = 'a option

type 'a m = 'a

type 'a e = 'a option

val lift : 'a m -> 'a t

lifts inner monad into the resulting monad

val run : 'a t -> 'a e

runs the computation

val void : 'a t -> unit t

void m computes m and discrards the result.

val sequence : unit t list -> unit t

sequence xs computes a sequence of computations xs in the left to right order.

val forever : 'a t -> 'b t

forever xs creates a computationt that never returns.

module Fn : sig ... end

Various function combinators lifted into the Kleisli category.

module Pair : sig ... end

The pair interface lifted into the monad.

module Triple : sig ... end

The triple interface lifted into a monad.

module Lift : sig ... end

Lifts functions into the monad.

module Exn : sig ... end

Interacting between monads and language exceptions

module Collection : sig ... end

Lifts collection interface into the monad.

module List : sig ... end

The Monad.Collection.S interface for lists

module Seq : sig ... end

The Monad.Collection.S interface for sequences

val (>=>) : ('a -> 'b t) -> ('b -> 'c t) -> 'a -> 'c t

f >=> g is fun x -> f x >>= g

val (!!) : 'a -> 'a t

!!x is return x

val (!$) : ('a -> 'b) -> 'a t -> 'b t

!$f is Lift.unary f

val (!$$) : ('a -> 'b -> 'c) -> 'a t -> 'b t -> 'c t

!$$f is Lift.binary f

val (!$$$) : ('a -> 'b -> 'c -> 'd) -> 'a t -> 'b t -> 'c t -> 'd t

!$$$f is Lift.ternary f

val (!$$$$) : ('a -> 'b -> 'c -> 'd -> 'e) -> 'a t -> 'b t -> 'c t -> 'd t -> 'e t

!$$$$f is Lift.quaternary f

val (!$$$$$) : ('a -> 'b -> 'c -> 'd -> 'e -> 'f) -> 'a t -> 'b t -> 'c t -> 'd t -> 'e t -> 'f t

!$$$$$f is Lift.quinary f

val (>>=) : 'a t -> ('a -> 'b t) -> 'b t

val (>>|) : 'a t -> ('a -> 'b) -> 'b t

module Monad_infix : sig ... end

val bind : 'a t -> f:('a -> 'b t) -> 'b t

val return : 'a -> 'a t

val map : 'a t -> f:('a -> 'b) -> 'b t

val join : 'a t t -> 'a t

val ignore_m : 'a t -> unit t

val all : 'a t list -> 'a list t

val all_unit : unit t list -> unit t

val all_ignore : unit t list -> unit t

module Let_syntax : sig ... end

module Syntax : sig ... end

Monadic operators, see Monad.Syntax.S for more.

val pure : 'a -> 'a t

pure x creates a computation that results in x.

val accept : 'a -> 'a t

accept x accepts x as a result of computation. (Same as pure x.

val reject : unit -> 'a t

reject () rejects the rest of computation sequence, and terminate the computation with the zero result (Same as zero ()

val guard : bool -> unit t

guard cond ensures cond is true in the rest of computation. Otherwise the rest of the computation is rejected.

val on : bool -> unit t -> unit t

on cond x computes x only iff cond is true

val unless : bool -> unit t -> unit t

unless cond x computes x unless cond is true.

val zero : unit -> 'a t

zero () constructs a zero element

val plus : 'a t -> 'a t -> 'a t

plus x y an associative operation.

type 'a error = unit

val fail : 'b error -> 'a t

fail err diverges the computation, possibly providing an extra information in a value of type _ error.

val catch : 'a t -> ('b error -> 'a t) -> 'a t

catch m f if m diverges with some bottom value err, the f err is a result of the whole computation, otherwise returns m.

module T1 : functor (M : Core_kernel.T1) -> sig ... end

module T2 : functor (M : Core_kernel.T2) -> sig ... end

module Make : functor (M : Monad) -> S with type 'a m := 'a T1(M).m and type 'a t := 'a T1(M).t and type 'a e := 'a T1(M).e

Make(M) composes the option monad with the monad M.

module Make2 : functor (M : Monad2) -> S2 with type ('a, 'e) t := ('a, 'e) T2(M).t and type ('a, 'e) m := ('a, 'e) T2(M).m and type ('a, 'e) e := ('a, 'e) T2(M).e

Make2(M) composes the option monad with the monad M.