Module State.Make2

Parameters

module M : Monad

Signature

include Trans.S1 with type ('a, 'e) t := ('a, 'e) T2(M).t with type 'a m := 'a T2(M).m with type ('a, 'e) e := ('a, 'e) T2(M).e

val lift : 'a T2(M).m -> ('a, 'e) T2(M).t

lifts inner monad into the resulting monad

val run : ('a, 'e) T2(M).t -> ('a, 'e) T2(M).e

runs the computation

include Monad2 with type ('a, 's) t := ('a, 's) T2(M).t

val void : ('a, 'e) T2(M).t -> (unit, 'e) T2(M).t

void m computes m and discrards the result.

val sequence : (unit, 'e) T2(M).t list -> (unit, 'e) T2(M).t

sequence xs computes a sequence of computations xs in the left to right order.

val forever : ('a, 'e) T2(M).t -> ('b, 'e) T2(M).t

forever xs creates a computationt that never returns.

module Fn : sig ... end

Various function combinators lifted into the Kleisli category.

module Pair : sig ... end

The pair interface lifted into the monad.

module Triple : sig ... end

The triple interface lifted into a monad.

module Lift : sig ... end

Lifts functions into the monad.

module Exn : sig ... end

Interacting between monads and language exceptions

module Collection : sig ... end

Lifts collection interface into the monad.

module List : Collection.S with type 'a t := 'a list

The Monad.Collection.S2 interface for lists

module Seq : Collection.S with type 'a t := 'a Core_kernel.Sequence.t

The Monad.Collection.S2 interface for sequences

include Syntax.S2 with type ('a, 'e) t := ('a, 'e) T2(M).t

val (>=>) : ('a -> ('b, 'e) T2(M).t) -> ('b -> ('c, 'e) T2(M).t) -> 'a -> ('c, 'e) T2(M).t

f >=> g is fun x -> f x >>= g

val (!!) : 'a -> ('a, 'e) T2(M).t

!!x is return x

val (!$) : ('a -> 'b) -> ('a, 'e) T2(M).t -> ('b, 'e) T2(M).t

!$f is Lift.unary f

val (!$$) : ('a -> 'b -> 'c) -> ('a, 'e) T2(M).t -> ('b, 'e) T2(M).t -> ('c, 'e) T2(M).t

!$$f is Lift.binary f

val (!$$$) : ('a -> 'b -> 'c -> 'd) -> ('a, 'e) T2(M).t -> ('b, 'e) T2(M).t -> ('c, 'e) T2(M).t -> ('d, 'e) T2(M).t

!$$$f is Lift.ternary f

val (!$$$$) : ('a -> 'b -> 'c -> 'd -> 'e) -> ('a, 's) T2(M).t -> ('b, 's) T2(M).t -> ('c, 's) T2(M).t -> ('d, 's) T2(M).t -> ('e, 's) T2(M).t

!$$$$f is Lift.quaternary f

val (!$$$$$) : ('a -> 'b -> 'c -> 'd -> 'e -> 'f) -> ('a, 's) T2(M).t -> ('b, 's) T2(M).t -> ('c, 's) T2(M).t -> ('d, 's) T2(M).t -> ('e, 's) T2(M).t -> ('f, 's) T2(M).t

!$$$$$f is Lift.quinary f

include Syntax.Let.S2 with type ('a, 'e) t := ('a, 'e) T2(M).t

val let* : ('a, 'e) T2(M).t -> ('a -> ('b, 'e) T2(M).t) -> ('b, 'e) T2(M).t

let* r = f x in b is f x >>= fun r -> b

val and* : ('a, 'e) T2(M).t -> ('b, 'e) T2(M).t -> ('a * 'b, 'e) T2(M).t

monoidal product

val let+ : ('a, 'e) T2(M).t -> ('a -> 'b) -> ('b, 'e) T2(M).t

let+ r = f x in b is f x >>| fun r -> b

val and+ : ('a, 'e) T2(M).t -> ('b, 'e) T2(M).t -> ('a * 'b, 'e) T2(M).t

monoidal product

include Core_kernel.Monad.S2 with type ('a, 'e) t := ('a, 'e) T2(M).t

val (>>=) : ('a, 'e) T2(M).t -> ('a -> ('b, 'e) T2(M).t) -> ('b, 'e) T2(M).t

val (>>|) : ('a, 'e) T2(M).t -> ('a -> 'b) -> ('b, 'e) T2(M).t

module Let_syntax : sig ... end

module Monad_infix : sig ... end

val bind : ('a, 'e) T2(M).t -> f:('a -> ('b, 'e) T2(M).t) -> ('b, 'e) T2(M).t

val return : 'a -> ('a, 'b) T2(M).t

val map : ('a, 'e) T2(M).t -> f:('a -> 'b) -> ('b, 'e) T2(M).t

val join : (('a, 'e) T2(M).t, 'e) T2(M).t -> ('a, 'e) T2(M).t

val ignore_m : ('a, 'e) T2(M).t -> (unit, 'e) T2(M).t

val all : ('a, 'e) T2(M).t list -> ('a list, 'e) T2(M).t

val all_unit : (unit, 'e) T2(M).t list -> (unit, 'e) T2(M).t

module Let : Syntax.Let.S2 with type ('a, 'e) t := ('a, 'e) T2(M).t

Monadic Binding Operators.

module Syntax : Syntax.S2 with type ('a, 'e) t := ('a, 'e) T2(M).t

Monadic operators, see Monad.Syntax.S2 for more.

val put : 's -> (unit, 's) T2(M).t

put s changes the current state to s

val get : unit -> ('s, 's) T2(M).t

get s gets the current state

val gets : ('s -> 'r) -> ('r, 's) T2(M).t

gets p projects the current state with the function p

val update : ('s -> 's) -> (unit, 's) T2(M).t

update f updates the current state with the function f