Module type Multi.S2

include Trans.S1

type ('a, 'e) t

type 'a m

type ('a, 'e) e

val lift : 'a m -> ('a, 'e) t

lifts inner monad into the resulting monad

val run : ('a, 'e) t -> ('a, 'e) e

runs the computation

type id

module Id : Core_kernel.Identifiable.S with type t = id

val global : id

the identifier of the global (initial) state.

val fork : unit -> (unit, 'e) t

fork () forks the current state.

val switch : id -> (unit, 'e) t

switch id switches to the state with the given id if such state is alive, otherwise switches to the closest alive ancestor of the state with the given id

val parent : unit -> (id, 'e) t

parent () returns an identifier of the closest alive parent.

val ancestor : id list -> (id, 'e) t

ancestor ids returns an identifier of the closest common ancestor of states with the given identifiers.

val current : unit -> (id, 'e) t

current id returns an identifier of current state.

val kill : id -> (unit, 'e) t

kill id kills a state with the specified id. If id corresponds to the current state, then switches to the closest ancestor. If id = global then do nothing.

val forks : unit -> (id Core_kernel.Sequence.t, 'e) t

forks xs returns a sequence of all alive states

val status : id -> (status, 'e) t

status id returns a status of a state with the given id

include S2 with type ('a, 'e) t := ('a, 'e) t and type ('a, 'e) e := ('a, 'e) e and type 'a m := 'a m

include Trans.S1

type ('a, 'e) t

type 'a m

type ('a, 'e) e

val lift : 'a m -> ('a, 'e) t

lifts inner monad into the resulting monad

val run : ('a, 'e) t -> ('a, 'e) e

runs the computation

include Monad2 with type ('a, 's) t := ('a, 's) t

type ('a, 'e) t

val void : ('a, 'e) t -> (unit, 'e) t

void m computes m and discrards the result.

val sequence : (unit, 'e) t list -> (unit, 'e) t

sequence xs computes a sequence of computations xs in the left to right order.

val forever : ('a, 'e) t -> ('b, 'e) t

forever xs creates a computationt that never returns.

module Fn : sig ... end

Various function combinators lifted into the Kleisli category.

module Pair : sig ... end

The pair interface lifted into the monad.

module Triple : sig ... end

The triple interface lifted into a monad.

module Lift : sig ... end

Lifts functions into the monad.

module Exn : sig ... end

Interacting between monads and language exceptions

module Collection : sig ... end

Lifts collection interface into the monad.

module List : Collection.S with type 'a t := 'a list

The Monad.Collection.S2 interface for lists

module Seq : Collection.S with type 'a t := 'a Core_kernel.Sequence.t

The Monad.Collection.S2 interface for sequences

include Syntax.S2 with type ('a, 'e) t := ('a, 'e) t

type ('a, 'e) t

val (>>=) : ('a, 'e) t -> ('a -> ('b, 'e) t) -> ('b, 'e) t

m >>= f is bind m f

val (>>|) : ('a, 'e) t -> ('a -> 'b) -> ('b, 'e) t

m >>= f is map m ~f

val (>=>) : ('a -> ('b, 'e) t) -> ('b -> ('c, 'e) t) -> 'a -> ('c, 'e) t

f >=> g is fun x -> f x >>= g

val (!!) : 'a -> ('a, 'e) t

!!x is return x

val (!$) : ('a -> 'b) -> ('a, 'e) t -> ('b, 'e) t

!$f is Lift.unary f

val (!$$) : ('a -> 'b -> 'c) -> ('a, 'e) t -> ('b, 'e) t -> ('c, 'e) t

!$$f is Lift.binary f

val (!$$$) : ('a -> 'b -> 'c -> 'd) -> ('a, 'e) t -> ('b, 'e) t -> ('c, 'e) t -> ('d, 'e) t

!$$$f is Lift.ternary f

val (!$$$$) : ('a -> 'b -> 'c -> 'd -> 'e) -> ('a, 's) t -> ('b, 's) t -> ('c, 's) t -> ('d, 's) t -> ('e, 's) t

!$$$$f is Lift.quaternary f

val (!$$$$$) : ('a -> 'b -> 'c -> 'd -> 'e -> 'f) -> ('a, 's) t -> ('b, 's) t -> ('c, 's) t -> ('d, 's) t -> ('e, 's) t -> ('f, 's) t

!$$$$$f is Lift.quinary f

include Core_kernel.Monad.S2 with type ('a, 'e) t := ('a, 'e) t

type ('a, 'e) t

include Base__.Monad_intf.Infix2 with type ('a, 'e) t := ('a, 'e) t

type ('a, 'e) t

val (>>=) : ('a, 'e) t -> ('a -> ('b, 'e) t) -> ('b, 'e) t

val (>>|) : ('a, 'e) t -> ('a -> 'b) -> ('b, 'e) t

include Base__.Monad_intf.Syntax2 with type ('a, 'e) t := ('a, 'e) t

type ('a, 'e) t

module Let_syntax : sig ... end

module Monad_infix : Base__.Monad_intf.Infix2 with type ('a, 'e) t := ('a, 'e) t

val bind : ('a, 'e) t -> f:('a -> ('b, 'e) t) -> ('b, 'e) t

val return : 'a -> ('a, _) t

val map : ('a, 'e) t -> f:('a -> 'b) -> ('b, 'e) t

val join : (('a, 'e) t, 'e) t -> ('a, 'e) t

val ignore_m : (_, 'e) t -> (unit, 'e) t

val all : ('a, 'e) t list -> ('a list, 'e) t

val all_unit : (unit, 'e) t list -> (unit, 'e) t

val all_ignore : (unit, 'e) t list -> (unit, 'e) t

module Syntax : Syntax.S2 with type ('a, 'e) t := ('a, 'e) t

Monadic operators, see Monad.Syntax.S2 for more.

val put : 's -> (unit, 's) t

put s changes the current state to s

val get : unit -> ('s, 's) t

get s gets the current state

val gets : ('s -> 'r) -> ('r, 's) t

gets p projects the current state with the function p

val update : ('s -> 's) -> (unit, 's) t

update f updates the current state with the function f