Module type Choice.S2

The binary choice monad interface

type ('a, 'e) t

include Basic2 with type ('a, 'e) t := ( 'a, 'e ) t

val pure : 'a -> ( 'a, 'e ) t

pure x creates a computation that results in x.

val zero : unit -> ( 'a, 'e ) t

zero () creates a computation that has no result.

val accept : 'a -> ( 'a, 'e ) t

accept x accepts x as a result of computation. (Same as pure x.

val reject : unit -> ( 'a, 'e ) t

reject () rejects the rest of computation sequence, and terminate the computation with the zero result (Same as zero ()

val guard : bool -> ( unit, 'e ) t

guard cond ensures cond is true in the rest of computation. Otherwise the rest of the computation is rejected.

val on : bool -> ( unit, 'e ) t -> ( unit, 'e ) t

on cond x computes x only iff cond is true

val unless : bool -> ( unit, 'e ) t -> ( unit, 'e ) t

unless cond x computes x unless cond is true.