Module Monad.Fail

A fail monad interface.

Fail monads are used to represent partial computation. There are several Fail monads with different flavors and kinds. This module provides a base abstraction for them, namely the fail and catch functions.

Implemented by:

Monad.Result.Make

Monad.Result.Make2

Monad.Result.Error

Monad.Result.Error.Make

Monad.Result.Exception

Monad.Result.Exception.Make

module type S = sig ... end

The unary fail monad interface.

module type S2 = sig ... end

The binary fail monad interface.