Module Syntax.Let

Monadic bindings operators.

This operators allows to write

let* r = computation x in body

instead of the old infix style

computation x >>= fun r -> body

The let* and and* operators stand for the monad part of the binding operators interface and let+ with and+ stand for the applicative part of the binding operators interface.

since

2.2.0 and OCaml 4.08.0

module type S = sig ... end

module type S2 = sig ... end